Rename to lowercase lexer
