+++ server/src/lexer.ts
